+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # datannur
 
+## unreleased
+
+- change : just a test for new release process
+
 ## 0.12.3 (2025-09-06)
 
 - fix : release url