+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - change : just a test for new release process
 - change : just another test for new release process
+- change : improve update_app.py script
 - remove : gitignore and gitattributes from public folder
 
 ## 0.12.3 (2025-09-06)