+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # datannur
 
-## unreleased
+## 0.12.4 (2025-09-07)
 
 - change : just a test for new release process
 - change : just another test for new release process
 - change : improve update_app.py script
 - change : refactor get_config() and  use some early return in update_app.py
+- change : refactor vite.config.ts
 - remove : gitignore and gitattributes from public folder
 - remove : svelte.config.js and move its content to vite.config.js